Replace 'signin' component from 'login' to 'auth' folder issues/#13
